refactor view structure
